# swagger.yaml
openapi: 3.0.0
info:
  description: This is the API docmentation for Telemed
  version: "1.0.0-oas3"
  title: Way-Farer API
  contact:
    email: mcdavidemereuwa95@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
servers:
  - url: 'http://localhost:1337/'
paths:
...
  /api/v1/trips/{tripId}:
    get:
      tags:
        - Trips
      summary: Get one trip
      parameters:
        - in: params
          name: tripId
          schema:
            type: number
            required: true
        - in: header
          name: x-access-token
          schema:
            type: string
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: status either 'success'
                  data:
                    type: object
                    properties:
                      trip_id:
                        type: number
                        description: id of created trip
                      bus_id:
                        type: number
                        description: id of bus to be used
                      origin:
                        type: string
                        description: take off location
                      destination:
                        type: string
                        description: trip destination
                      trip_date:
                        type: string
                        description: date of trip
                      fare:
                        type: float
                        description: cost of trip
                    description: data from request
                example:
                  status: 'success'
                  data:
                    trip_id: 1
                    bus_id: 1
                    origin: egbeda
                    destination: oshodi
                    fare: 150.00
                    trip_date: timestamp

          description: List of trips
        '404':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  error:
                    type: string
                example:
                  status: 'error'
                  error: 'tripId does not exists'
          description: an invalid email
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  error:
                    type: string
                example:
                  status: 'error'
                  error: 'you are not authorized to access this endpoint'
          description: Invalid Authorization or jwt token
        '500':
          description: Internal Server Error
    patch:
      tags:
        - Trips
        - Admin
      summary: cancels a trip
      operationId: canceltrip
      description: |
        cancels a trip
      parameters:
        - in: params
          name: tripId
          schema:
            type: string
            required: true
        - in: header
          name: x-access-token
          schema:
            type: string
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: status either 'success'
                  data:
                    type: object
                    properties:
                      messsage:
                        type: string
                example:
                  status: 'success'
                  data:
                    message: Trip successfully cancelled
          description: Trip successfully cancelled
        '401':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  error:
                    type: string
                example:
                  status: 'error'
                  error: 'tripId is required'
          description: 'invalid input, object invalid'
        '404':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  error:
                    type: string
                example:
                  status: 'error'
                  error: 'tripId is does not exists'
          description: an invalid email
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  error:
                    type: string
                example:
                  status: 'error'
                  error: 'you are not authorized to access this endpoint'
          description: Invalid Authorization or jwt token
        '500':
          description: Internal Server Error
...
  /api/v1/bookings/{bookingId}:
    delete:
      tags:
        - Bookings
      summary: books a trip
      operationId: deleteBooking
      description: |
        Books a trip for the user and reserves a seat
      parameters:
        - in: params
          name: bookingId
          schema:
            type: number
          required: true
          description: id of booking to delete
        - in: header
          name: x-access-token
          schema:
            type: string
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: status either 'success'
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                example:
                  status: 'success'
                  data:
                    message: 'booking deleted successfully'
          description: Booking successfully created
        '401':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  error:
                    type: string
                example:
                  status: 'error'
                  error: 'bookingId is required'
          description: 'invalid input, object invalid'
        '404':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  error:
                    type: string
                example:
                  status: 'error'
                  error: 'bookingId is does not exists'
          description: an invalid email
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  error:
                    type: string
                example:
                  status: 'error'
                  error: 'you are not authorized to access this endpoint'
          description: Invalid Authorization or jwt token
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                trip_id: 1 